Split HOC with the main component
