+++ src/components/screens/lockedBalance/transactionStatus/transactionStatus.js
@@ -1,0 +1,43 @@
+import React, { useEffect } from 'react';
+import { routes } from '@constants';
+import { PrimaryButton } from '@toolbox/buttons';
+import { TransactionResult, getBroadcastStatus } from '@shared/transactionResult';
+import statusMessages from './statusMessages';
+import styles from './status.css';
+
+const TransactionStatus = (props) => {
+  const {
+    t, history, transactionInfo, transactions, transactionBroadcasted,
+  } = props;
+  const status = getBroadcastStatus(transactions, false); // handle HW error
+  const onSuccess = () => history.push(routes.wallet.path);
+  const template = statusMessages(t, onSuccess)[status.code];
+
+  useEffect(() => {
+    if (transactionInfo) transactionBroadcasted(transactionInfo);
+  }, [transactionInfo]);
+
+  return (
+    <div className={`${styles.wrapper} transaction-status`}>
+      <TransactionResult
+        t={t}
+        illustration="default"
+        status={status}
+        title={template.title}
+        message={template.message}
+        className={styles.content}
+      >
+        {template.button && (
+          <PrimaryButton
+            onClick={template.button.onClick}
+            className={`${template.button.className} dialog-close-button`}
+          >
+            {template.button.title}
+          </PrimaryButton>
+        )}
+      </TransactionResult>
+    </div>
+  );
+};
+
+export default TransactionStatus;
